Reorganize common modules, typedocs
