+++ frontend/src/app/_directives/title.directive.ts
@@ -14,10 +14,11 @@ export class TitleDirective implements OnInit {
   type: string;
 
   constructor(private router: Router, private activatedRoute: ActivatedRoute, private titleService: Title) {}
-  ngOnInit() {
-    // Look at the current route to see if a title is set,
-    // if so, set the page title via titleService
 
+  /**
+   * Look at the current route to see if a title is set, if so, set the page title via titleService
+   */
+  ngOnInit() {
     this.router.events
       .filter(event => event instanceof NavigationEnd)
       .map(() => this.activatedRoute)
