+++ frontend/src/app/_pipes/base64.pipe.ts
@@ -4,6 +4,10 @@ import { Pipe, PipeTransform } from '@angular/core';
   name: 'base64'
 })
 export class Base64 implements PipeTransform {
+
+  /**
+   * @returns base64 encoded string
+   */
   public transform(value: any) {
     return btoa(value);
   }
