+++ frontend/src/app/_pipes/camel-to-hyphens.pipe.ts
@@ -2,6 +2,10 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({ name: 'CamelToHyphens' })
 export class CamelToHyphensPipe implements PipeTransform {
+
+  /**
+   * @returns converts camel case to hyphen case
+   */
   transform(value: string): string {
     if (value) {
       return value
