+++ frontend/src/app/_pipes/filter.pipe.ts
@@ -4,6 +4,10 @@ import { Pipe, PipeTransform } from '@angular/core';
   name: 'Filter'
 })
 export class FilterPipe implements PipeTransform {
+
+  /**
+   * @returns Filtered array
+   */
   transform(items: any, field: string, args: string): any {
     if (items) {
       return items.filter(item => item[field] === args);
