+++ frontend/src/app/app.component.html
@@ -7,7 +7,7 @@
   </div>
 </header>
 <main id="main-content" tabindex="0">
-  <app-status [message]="status.message" [heading]="status.heading"></app-status>
+  <app-alert [message]="status.message" [heading]="status.heading"></app-alert>
   <router-outlet></router-outlet>
 </main>
 <app-page-footer [version]="version" [buildDate]="buildDate"></app-page-footer>
