+++ frontend/src/app/application-forms/application-submitted/application-submitted.component.ts
@@ -1,4 +1,3 @@
-import { SpecialUseApplication } from '../../_models/special-use-application';
 import { ApplicationService } from '../../_services/application.service';
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
