+++ frontend/src/app/shared/footer/footer.component.ts
